+++ dev/tests/spec/match.spec.js
@@ -29,18 +29,16 @@ describe('Match', function(){
     });
 
     it('should match params', function(){
-
         var s = crossroads.addRoute('/{foo}');
 
         expect( s.match('/lorem-ipsum') ).toBe( true );
         expect( s.match('/lorem-ipsum/') ).toBe( true );
         expect( s.match('/lorem-ipsum/dolor') ).toBe( false );
-        expect( s.match('lorem-ipsum') ).toBe( false );
+        expect( s.match('lorem-ipsum') ).toBe( true );
         expect( s.match('/123') ).toBe( true );
         expect( s.match('/123/') ).toBe( true );
-        expect( s.match('123') ).toBe( false );
+        expect( s.match('123') ).toBe( true );
         expect( s.match('123/45') ).toBe( false );
-
     });
 
     it('should match optional params', function(){
@@ -48,8 +46,7 @@ describe('Match', function(){
         expect( s.match('lorem-ipsum') ).toBe( true );
         expect( s.match('') ).toBe( true );
         expect( s.match('lorem-ipsum/dolor') ).toBe( false );
-        expect( s.match('/lorem-ipsum/') ).toBe( false );
-
+        expect( s.match('/lorem-ipsum/') ).toBe( true );
     });
 
     it('should match normal params and optional params', function(){
@@ -57,7 +54,7 @@ describe('Match', function(){
         expect( s.match('/lorem-ipsum') ).toBe( true );
         expect( s.match('/lorem-ipsum/') ).toBe( true );
         expect( s.match('/lorem-ipsum/dolor') ).toBe( true );
-        expect( s.match('123/45') ).toBe( false );
+        expect( s.match('123/45') ).toBe( true );
     });
 
     it('should work even with optional params on the middle of pattern', function(){
@@ -112,6 +109,99 @@ describe('Match', function(){
             expect( s.match('/foo/asd/edit/qwe') ).toBe( false );
         });
     });
+
+    describe('query string', function () {
+        it('should match query string as first segment', function () {
+            var r = crossroads.addRoute('{?q}');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( false );
+            expect( r.match('/foo') ).toBe( false );
+            expect( r.match('foo/') ).toBe( false );
+            expect( r.match('/foo/') ).toBe( false );
+            expect( r.match('?foo') ).toBe( true );
+            expect( r.match('?foo=bar') ).toBe( true );
+            expect( r.match('?foo=bar&lorem=123') ).toBe( true );
+        });
+
+        it('should match optional query string as first segment', function () {
+            var r = crossroads.addRoute(':?q:');
+            expect( r.match('') ).toBe( true );
+            expect( r.match('foo') ).toBe( false );
+            expect( r.match('/foo') ).toBe( false );
+            expect( r.match('foo/') ).toBe( false );
+            expect( r.match('/foo/') ).toBe( false );
+            expect( r.match('?foo') ).toBe( true );
+            expect( r.match('?foo=bar') ).toBe( true );
+            expect( r.match('?foo=bar&lorem=123') ).toBe( true );
+        });
+
+        it('should match query string as 2nd segment', function () {
+            var r = crossroads.addRoute('{a}{?q}');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( false );
+            expect( r.match('/foo') ).toBe( false );
+            expect( r.match('foo/') ).toBe( false );
+            expect( r.match('/foo/') ).toBe( false );
+            expect( r.match('foo?foo') ).toBe( true );
+            expect( r.match('foo?foo=bar') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123') ).toBe( true );
+        });
+
+        it('should match optional query string as 2nd segment', function () {
+            var r = crossroads.addRoute('{a}:?q:');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( true );
+            expect( r.match('/foo') ).toBe( true );
+            expect( r.match('foo/') ).toBe( true );
+            expect( r.match('/foo/') ).toBe( true );
+            expect( r.match('foo?foo') ).toBe( true );
+            expect( r.match('foo?foo=bar') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123') ).toBe( true );
+        });
+
+        it('should match query string as middle segment', function () {
+            //if hash is required should use the literal "#" to avoid matching
+            //the last char of string as a string "foo?foo" shouldn't match
+            var r = crossroads.addRoute('{a}{?q}#{hash}');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( false );
+            expect( r.match('/foo') ).toBe( false );
+            expect( r.match('foo/') ).toBe( false );
+            expect( r.match('/foo/') ).toBe( false );
+            expect( r.match('foo?foo') ).toBe( false );
+            expect( r.match('foo?foo#bar') ).toBe( true );
+            expect( r.match('foo?foo=bar#bar') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123#bar') ).toBe( true );
+        });
+
+        it('should match optional query string as middle segment', function () {
+            var r = crossroads.addRoute('{a}:?q::hash:');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( true );
+            expect( r.match('/foo') ).toBe( true );
+            expect( r.match('foo/') ).toBe( true );
+            expect( r.match('/foo/') ).toBe( true );
+            expect( r.match('foo?foo') ).toBe( true );
+            expect( r.match('foo?foo=bar') ).toBe( true );
+            expect( r.match('foo?foo=bar#bar') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123#bar') ).toBe( true );
+        });
+
+        it('should match query string even if not using the special query syntax', function () {
+            var r = crossroads.addRoute('{a}?{q}#{hash}');
+            expect( r.match('') ).toBe( false );
+            expect( r.match('foo') ).toBe( false );
+            expect( r.match('/foo') ).toBe( false );
+            expect( r.match('foo/') ).toBe( false );
+            expect( r.match('/foo/') ).toBe( false );
+            expect( r.match('foo?foo') ).toBe( false );
+            expect( r.match('foo?foo#bar') ).toBe( true );
+            expect( r.match('foo?foo=bar#bar') ).toBe( true );
+            expect( r.match('foo?foo=bar&lorem=123#bar') ).toBe( true );
+        });
+    });
+
 
     describe('slash between params are optional', function(){
 
@@ -232,6 +322,108 @@ describe('Match', function(){
             expect( a.match('/123asd/45') ).toBe( true );
             expect( a.match('/123asd/45/') ).toBe( true );
             expect( a.match('/123asd/45/qwe') ).toBe( false );
+        });
+
+    });
+
+
+    describe('strict slash rules', function () {
+
+        afterEach(function(){
+            crossroads.patternLexer.loose();
+        });
+
+        it('should only match if traling slashes match the original pattern', function () {
+            crossroads.patternLexer.strict();
+
+            var a = crossroads.addRoute('{foo}');
+            expect( a.match('foo') ).toBe( true );
+            expect( a.match('/foo') ).toBe( false );
+            expect( a.match('foo/') ).toBe( false );
+            expect( a.match('/foo/') ).toBe( false );
+
+            var b = crossroads.addRoute('/{foo}');
+            expect( b.match('foo') ).toBe( false );
+            expect( b.match('/foo') ).toBe( true );
+            expect( b.match('foo/') ).toBe( false );
+            expect( b.match('/foo/') ).toBe( false );
+
+            var c = crossroads.addRoute('');
+            expect( c.match() ).toBe( true );
+            expect( c.match('') ).toBe( true );
+            expect( c.match('/') ).toBe( false );
+            expect( c.match('foo') ).toBe( false );
+
+            var d = crossroads.addRoute('/');
+            expect( d.match() ).toBe( false );
+            expect( d.match('') ).toBe( false );
+            expect( d.match('/') ).toBe( true );
+            expect( d.match('foo') ).toBe( false );
+        });
+
+    });
+
+
+    describe('loose slash rules', function () {
+
+        beforeEach(function(){
+            crossroads.patternLexer.loose();
+        });
+
+        it('should treat single slash and empty string as same', function () {
+            var c = crossroads.addRoute('');
+            expect( c.match() ).toBe( true );
+            expect( c.match('') ).toBe( true );
+            expect( c.match('/') ).toBe( true );
+            expect( c.match('foo') ).toBe( false );
+
+            var d = crossroads.addRoute('/');
+            expect( d.match() ).toBe( true );
+            expect( d.match('') ).toBe( true );
+            expect( d.match('/') ).toBe( true );
+            expect( d.match('foo') ).toBe( false );
+        });
+
+    });
+
+    describe('legacy slash rules', function () {
+
+        beforeEach(function(){
+            crossroads.patternLexer.legacy();
+        });
+
+        afterEach(function(){
+            crossroads.patternLexer.loose();
+        });
+
+        it('should treat single slash and empty string as same', function () {
+            var c = crossroads.addRoute('');
+            expect( c.match() ).toBe( true );
+            expect( c.match('') ).toBe( true );
+            expect( c.match('/') ).toBe( true );
+            expect( c.match('foo') ).toBe( false );
+
+            var d = crossroads.addRoute('/');
+            expect( d.match() ).toBe( true );
+            expect( d.match('') ).toBe( true );
+            expect( d.match('/') ).toBe( true );
+            expect( d.match('foo') ).toBe( false );
+        });
+
+        it('slash at end of string is optional', function () {
+            var a = crossroads.addRoute('/foo');
+            expect( a.match('/foo') ).toEqual( true );
+            expect( a.match('/foo/') ).toEqual( true );
+            expect( a.match('/foo/bar') ).toEqual( false );
+        });
+
+        it('slash at begin of string is required', function () {
+            var a = crossroads.addRoute('/foo');
+            expect( a.match('/foo') ).toEqual( true );
+            expect( a.match('/foo/') ).toEqual( true );
+            expect( a.match('foo') ).toEqual( false );
+            expect( a.match('foo/') ).toEqual( false );
+            expect( a.match('/foo/bar') ).toEqual( false );
         });
 
     });
@@ -365,6 +557,43 @@ describe('Match', function(){
                 expect( s.match('/123/45/67') ).toBe( true );
 
                 crossroads.shouldTypecast = prevTypecast; //restore
+            });
+
+        });
+
+
+        describe('query string', function () {
+
+            it('should validate with array', function () {
+                var r = crossroads.addRoute('/foo.php{?query}');
+                r.rules = {
+                    '?query' : ['lorem=ipsum&dolor=456', 'amet=789']
+                };
+                expect( r.match('foo.php?bar=123&ipsum=dolor') ).toBe( false );
+                expect( r.match('foo.php?lorem=ipsum&dolor=456') ).toBe( true );
+                expect( r.match('foo.php?amet=789') ).toBe( true );
+            });
+
+            it('should validate with RegExp', function () {
+                var r = crossroads.addRoute('/foo.php{?query}');
+                r.rules = {
+                    '?query' : /^lorem=\w+&dolor=\d+$/
+                };
+                expect( r.match('foo.php?bar=123&ipsum=dolor') ).toBe( false );
+                expect( r.match('foo.php?lorem=ipsum&dolor=12345') ).toBe( true );
+                expect( r.match('foo.php?lorem=ipsum&dolor=amet') ).toBe( false );
+            });
+
+            it('should validate with Function', function () {
+                var r = crossroads.addRoute('/foo.php{?query}');
+                r.rules = {
+                    '?query' : function(val, req, vals){
+                        return (val.lorem === 'ipsum' && typeof val.dolor === 'number');
+                    }
+                };
+                expect( r.match('foo.php?bar=123&ipsum=dolor') ).toBe( false );
+                expect( r.match('foo.php?lorem=ipsum&dolor=12345') ).toBe( true );
+                expect( r.match('foo.php?lorem=ipsum&dolor=amet') ).toBe( false );
             });
 
         });